Producte

+++ src/Producte.java
@@ -2,8 +2,9 @@
 public class Producte {
     private String  nom;
     private float preu;
-    private int quantitat;
+    private float quantitat;
     private String idProducte;
+
 
     public Producte(String nom, float preu, int quantitat, String idProducte) {
         this.nom = nom;
